move evaluate_coco out of predictor
